pyproject.toml
[project]
name = "tigrex-plush"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "discord-py>=2.6.3",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "google-genai>=1.50.1",
    "gtts>=2.5.4",
    "keep-alive>=0.0.1",
    "openai>=2.7.2",
    "pynacl>=1.6.0",
    "pyttsx3>=2.99",
    "requests>=2.32.5",
    "yt-dlp>=2025.11.12",
]
